lab4: Split into modules
